Lowered the precision on courseOverGround, speedOverGround et al.

+++ service/src/main/java/eu/europa/ec/fisheries/uvms/movement/service/bean/SegmentBean.java
@@ -87,7 +87,7 @@ public class SegmentBean {
         segment.setDistance(positionCalculations.getDistanceBetweenPoints());
         segment.setSpeedOverGround(positionCalculations.getAvgSpeed());
         segment.setCourseOverGround(positionCalculations.getCourse());
-        segment.setDuration(positionCalculations.getDurationBetweenPoints());
+        segment.setDuration((int)positionCalculations.getDurationBetweenPoints());
 
         segment.setFromMovement(fromMovement);
         segment.setToMovement(toMovement);
